move onslope code to self-contained file
